Update index.html
Changing up some code so the styling looks better.  Want the images to show up before the search function on mobile devices.  Didn't like the way the container looked on mobile, want it to take up the whole screen.  Didn't like how small the map was so increased its size.